python 3 change

+++ writing_database.py
@@ -3,14 +3,15 @@ from pymongo import MongoClient
 import re
 import time
 from bs4 import BeautifulSoup
+import random
 # from selenium import webdriver
 # from selenium.webdriver.common.keys import Keys
 import urllib
 import sys
 
 
-def get_urls():
-    years = ["2017", "2018"]
+def get_urls(reset = False):
+    years = ["2018"]
     months = ["jan", "feb", "mar", "apr", "may", "jun", "jul", "aug", "sep", "oct", "nov", "dec"]
     dates = [str(n) for n in range(1,28)]
 
@@ -28,7 +29,9 @@ def get_urls():
     db = client["guardian"] #This is the name of the database
     urls = db["urls"] # this is the table in that database
 
-    result = db.urls.delete_many({}) # A fresh start to the DB table -> removing all entries
+    if reset:
+        result = db.urls.delete_many({}) # A fresh start to the DB table -> removing all entries
+        result = db.data.delete_many({})
     all_links = []
 
     for year in years:
@@ -191,37 +194,43 @@ def article_topics_title(url):
 
 
 
-def real_download():
+def real_download(title = 1, article = 1, topics = 1, comments_data = 0):
     client = MongoClient()
     db = client["guardian"] #This is the name of the database
     urls = db["urls"] # this is the table in that database
+    data = db["data"]
+
+    i = 0
 
     cursor = urls.find({}, no_cursor_timeout= True)
     for document in cursor:
+        seconds = random.randint(5, 39)
+        time.sleep(seconds)
         url = str(document['url'])
-        time.sleep(60)
         id_n = document['_id']
-        print ("\n")
-        print ("Working on the link -", url)
-
-
-#         soup = download_comments(url)
-    ###      comment_text_list, comment_id_list, author_id_list, auth_name_lst, upvotes_count_list
-#         comment_text_list, comment_id_list, author_id_list, auth_name_lst, upvotes_count_list = getting_comment_data(soup)
-
-        topics_list, article, title = article_topics_title(url)
+        topic_list, article, title = article_topics_title(url)
         print ("Title of the article is: ", title)
+        print ("\n")
 
-        urls.updateOne({'url': url},
-                        {"$set": {'title': title, 'article' : article,
-                        'topics_list': topics_list}}, upsert=True)
-#                         {"$set": {'title': title, 'article' : article, 'topics_list': topics_list,
-#                                 'comment_ids' : comment_id_list, 'comment_text' : comment_text_list,
-#                                 'author_ids' : author_id_list, 'author_name' : auth_name_lst,
-#                                 'upvotes' : upvotes_count_list}}, upsert=True)
+        if comments_data == 0 :
+            res = data.insert_one({'id' : i, 'url' : url, 'article' : article,
+                       'title' : title, 'topic_list' : topic_list})
+
+        if comments_data == 1:
+            soup = download_comments(url)
+            comment_text_list, comment_id_list, author_id_list, auth_name_lst, upvotes_count_list = getting_comment_data(soup)
+        ##      comment_text_list, comment_id_list, author_id_list, auth_name_lst, upvotes_count_list
+            res = data.insert_one({'id' : i, 'url': url, 'title': title, 'article' : article,
+                            'topics_list': topics_list, 'comment_ids' : comment_id_list,
+                            'comment_text' : comment_text_list,
+                            'author_ids' : author_id_list, 'author_name' : auth_name_lst,
+                            'upvotes' : upvotes_count_list})
+        i += 1
+        if i % 2 == 0:
+            print ("Message : {:d} documents downloaded.".format(i))
 
     return
 
 if __name__ == "__main__":
-    get_urls()
+    get_urls(reset = True)
     real_download()
